+++ main.go
@@ -3,5 +3,5 @@ package main
 import app "go-api/src/main"
 
 func main() {
-	app.StartApp()
+	app.Start()
 }